Refatorando código para deixar mais organizado
Só que eu não gostei do resultado final...
