+++ app/src/main/java/br/ufg/pos/fswmd/myfriendsapp/addfriend/activity/AddFriendActivity.kt
@@ -16,6 +16,7 @@ import android.view.MenuItem
 import android.widget.Toast
 import br.ufg.pos.fswmd.myfriendsapp.R
 import br.ufg.pos.fswmd.myfriendsapp.addfriend.listener.SaveButtonListener
+import br.ufg.pos.fswmd.myfriendsapp.addfriend.listener.TakePictureButtonListener
 import br.ufg.pos.fswmd.myfriendsapp.repository.FriendDatabaseHandler
 import br.ufg.pos.fswmd.myfriendsapp.repository.FriendRepository
 import kotlinx.android.synthetic.main.activity_main.*
@@ -26,12 +27,9 @@ import java.util.*
 class AddFriendActivity : AppCompatActivity(), AddFriendView {
 
     private val REQUEST_PERMISSION_CODE: Int = 0
-    private val REQUEST_PICTURE_CODE: Int = 1
 
     private var presenter: AddFriendPresenter? = null
     private var friendRepository: FriendRepository? = null
-
-    private var file: Uri? = null
 
     private var saveButtonListener: SaveButtonListener? = null
 
@@ -44,11 +42,25 @@ class AddFriendActivity : AppCompatActivity(), AddFriendView {
         presenter = AddFriendPresenterImpl(this, friendRepository!!)
         saveButtonListener = SaveButtonListener(presenter!!, name_edit_id, nickname_edit_id, description_edit_id)
 
+        // requesting permission to access camera and write in directory
+        checkAndRequestPermissions()
+
         save_button_id.setOnClickListener(
                 saveButtonListener
         )
 
-        // requesting permission to access camera and write in directory
+        if (take_picture_button_id.isEnabled) {
+            take_picture_button_id.setOnClickListener(
+                    TakePictureButtonListener(this)
+            )
+        }
+
+        supportActionBar!!.setDisplayHomeAsUpEnabled(true)
+        supportActionBar!!.setHomeButtonEnabled(true)
+        supportActionBar!!.setTitle(R.string.new_friend_title)
+    }
+
+    private fun checkAndRequestPermissions() {
         if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA)
                 != PackageManager.PERMISSION_GRANTED) {
             take_picture_button_id.isEnabled = false
@@ -56,46 +68,15 @@ class AddFriendActivity : AppCompatActivity(), AddFriendView {
                     arrayOf(Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE),
                     REQUEST_PERMISSION_CODE)
         }
-
-        if (take_picture_button_id.isEnabled) {
-            take_picture_button_id.setOnClickListener {
-                var takePictureIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-                if (getOutputMediaFile() != null) {
-                    file = FileProvider.getUriForFile(this, "${this.packageName}.provider", getOutputMediaFile()!!)
-                    takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, file)
-
-                    startActivityForResult(takePictureIntent, REQUEST_PICTURE_CODE)
-                }
-            }
-        }
-
-        // enabling button back on action bar
-        supportActionBar!!.setDisplayHomeAsUpEnabled(true)
-        supportActionBar!!.setHomeButtonEnabled(true)
-        supportActionBar!!.setTitle(R.string.new_friend_title)
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode.equals(REQUEST_PICTURE_CODE)
+        if (requestCode.equals(TakePictureButtonListener.REQUEST_PICTURE_CODE)
                         .and(resultCode.equals(Activity.RESULT_OK))) {
-            take_picture_button_id.setImageURI(file)
-            saveButtonListener!!.photoUrl = file.toString()
+            take_picture_button_id.setImageURI(TakePictureButtonListener.file)
+            saveButtonListener!!.photoUrl = TakePictureButtonListener.file?.toString()
         }
 
-    }
-
-    private fun getOutputMediaFile(): File? {
-        var mediaStorageDir = File(getExternalFilesDir(
-                Environment.DIRECTORY_PICTURES), "CameraDemo")
-
-        if (mediaStorageDir.exists().not()) {
-            if (mediaStorageDir.mkdirs().not()) {
-                return null
-            }
-        }
-
-        var timestamp = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
-        return File("${mediaStorageDir}${File.separator}IMG_${timestamp}_.mySecret")
     }
 
     override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
