+++ app/src/main/res/layout/activity_detail_friend.xml
@@ -14,6 +14,8 @@
         android:layout_centerHorizontal="true"
         android:layout_margin="8dp"
         android:contentDescription="@string/photo_friend_description"
+        android:adjustViewBounds="true"
+        android:scaleType="centerCrop"
         app:srcCompat="@drawable/image_default" />
 
     <TextView
