+++ app/src/main/res/layout/friend_row_list.xml
@@ -29,6 +29,8 @@
                 android:layout_marginRight="2dp"
                 android:layout_marginVertical="2dp"
                 android:contentDescription="@string/list_photo_friend_description"
+                android:adjustViewBounds="true"
+                android:scaleType="centerCrop"
                 app:srcCompat="@drawable/image_default" />
 
             <TextView
